🤖 refactor: cookiekeys.ts 위치 변경
