+++ src/app/volunteer/register/[...slug]/RegisterComplete.tsx
@@ -1,6 +1,6 @@
 import RegisterComplete from '@/components/volunteer/RegisterComplete/RegisterComplete';
 
-export default function Complete() {
+export default function RegisterCompleteServerComp() {
   return (
     <>
       <RegisterComplete />
