+++ src/app/volunteer/register/[...slug]/page.tsx
@@ -1,7 +1,7 @@
-import { COOKIE_REGISTER_EMAIL_KEY } from '@/api/cookieKeys';
 import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
 import RegisterMain from './RegisterMain';
+import { COOKIE_REGISTER_EMAIL_KEY } from '@/constants/cookieKeys';
 export default function Register({}) {
   // TODO
   // TOU 체크한 부분 서버로 전달하는 로직 필요
