+++ src/components/volunteer/RegisterComplete/RegisterComplete.tsx
@@ -10,7 +10,7 @@ import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import useHeader from '@/hooks/useHeader';
 import Cookies from 'js-cookie';
-import { COOKIE_REGISTER_EMAIL_KEY } from '@/api/cookieKeys';
+import { COOKIE_REGISTER_EMAIL_KEY } from '@/constants/cookieKeys';
 
 export default function RegisterComplete() {
   const [isOpened, openDialog, closeDialog] = useBooleanState(true);
