+++ src/utils/ky/hooks/afterResponse.ts
@@ -3,7 +3,7 @@ import Cookies from 'js-cookie';
 import {
   COOKIE_ACCESS_TOKEN_KEY,
   COOKIE_REFRESH_TOKEN_KEY
-} from '@/api/cookieKeys';
+} from '@/constants/cookieKeys';
 import { fetchRefresh } from '@/api/auth/volunteer/refresh';
 import { UNREGISTERED } from '@/api/authErrorCode';
 import { ApiErrorResponse } from '@/types/apiTypes';
